changing controllers names, commenting unused lines in postMockUsers
